hirshfeld_moderator.py, bug-fix gcc in Windows
The compilation of diff_finger.c in Windows with gcc yields an executable
with the file extension of ".exe" -- even if this was not explicitly
stated in the Python script.  Of course, this prevented its transfer to
the data, its work to compute the difference maps, and eventual space
cleaning.

As a rebase of branch gcc_fix in branch gcc_fix_rebase, this issue is
resolved.

+++ python_testing/hirshfeld_moderator.py
@@ -4,7 +4,7 @@
 # author:  nbehrnd@yahoo.com
 # license: GPL version 2
 # date:    2019-11-14 (YYYY-MM-DD)
-# edit:    2020-03-07 (YYYY-MM-DD)
+# edit:    2020-03-12 (YYYY-MM-DD)
 #
 """ This is a moderator script for Python to either moderate a Difference
 Hirshfeld surface analysis introduced with
@@ -275,8 +275,11 @@ def fingerprint_fortran():
                 entry, dat_file))
 
         sub.call(normalize, shell=True)
-
-    os.remove("fingerprint.x")
+    if platform.system().startswith("Linux"):
+        try:
+            os.remove("fingerprint.x")
+        except:
+            pass
     os.chdir(root)
     print("\nNormalization of .cxs files is completed.")
 
@@ -315,17 +318,28 @@ def compile_c():
 
 def shuttle_c():
     """ Shuttle the executable of diff_finger.c to the data. """
-    try:
-        shutil.copy("diff_finger", "cxs_workshop")
-    except IOError:
-        print("Error while copying C executable to the data.")
-        print("Check the presence of folder 'cxs_workshop'.")
-        sys.exit(0)
+    # clause for Linux-based computers:
+    if platform.system().startswith("Linux"):
+        try:
+            shutil.copy("diff_finger", "cxs_workshop")
+        except IOError:
+            print("Error while copying C executable to the data.")
+            print("Check the presence of folder 'cxs_workshop'.")
+            sys.exit(0)
 
-    try:
-        os.remove("diff_finger")  # space cleaning, root folder.
-    except IOError:
-        pass
+        try:
+            os.remove("diff_finger")  # space cleaning, root folder.
+        except IOError:
+            pass
+
+    if platform.system().startswith("Windows"):
+        try:
+            shutil.copy("diff_finger.exe", "cxs_workshop")
+        except IOError:
+            print("Error while copying C executable to the data.")
+            print("Check the presence of folder 'cxs_workshop'.")
+            sys.exit(0)
+
 
 
 def difference_maps_c():
@@ -348,8 +362,12 @@ def difference_maps_c():
 
             print("{} vs. {} to yield {}".format(reference_map, test_map,
                                                  difference_map))
-            difference_test = str("./diff_finger {} {} > {}".format(
-                reference_map, test_map, difference_map))
+            if platform.system().startswith("Linux"):
+                difference_test = str("./diff_finger {} {} > {}".format(
+                    reference_map, test_map, difference_map))
+            if platform.system().startswith("Windows"):
+                difference_test = str("diff_finger.exe {} {} > {}".format(
+                    reference_map, test_map, difference_map))
             try:
                 sub.call(difference_test, shell=True)
             except IOError:
@@ -357,8 +375,16 @@ def difference_maps_c():
 
         del fingerprint_register[0]
     print("\nComputation of difference maps is completed.")
-
-    os.remove("diff_finger")
+    if platform.system().startswith("Linux"):
+        try:
+            os.remove("diff_finger")
+        except:
+            pass
+    if platform.system().startswith("Windows"):
+        try:
+            os.remove("diff_finger.exe")
+        except:
+            pass
 
 
 def difference_maps_python():
@@ -373,7 +399,7 @@ def difference_maps_python():
     diff_register.sort()
 
     # compare the normalized 2D Hirshfeld surface maps
-    print("Fingerprint computation with Python:\n")
+    print("\nComputation of difference maps (Python) starts:")
     while len(diff_register) > 1:
         for entry in diff_register[1:]:
             reference_file = diff_register[0]
